Implementing middle-level function id_cust(), and two low-level functions; pnt_id_cust(), and get_addr(), which are called from id_cust().
id_cust() prompts for customer identification information. We've
implemented it in the same file as id_user() to allow order data in
memory to be shared by them, but not by other functions linked into the
program. We've also defined that data in SEMIGLOBAL variables.
pnt_id_cust() simply paints the field names used to prompt the user for
customer identification. get_addr() prompts the user for the fields that
make up a street address.