Implementing middle-level function ord_itms(), and a stub function inv_find() called from ord_itms().
ord_itms() prompts for which and how many products the identified
customer is ordering, while displaying the data as a spreadsheet,
updating the total price with each input. inv_find() is supposed to look
up a part's data in the inventory and return true or false, as well as
pass back the description, price, and weight; but it will be replaced by
a operational version further on.