Created the program INVMAINT to build and maintain the inventory-file data.
The program INVMAINT will build the inventory-file data. Its goal is to
list and add parts to the inventory. It will change the part-numbering
scheme so that each part in the inventory is assigned an integer as its
part number. Thus, the position of the record in the file will be the
part number.